- debian packages for python2 and python3 - replaces standalone script by module with entrypoint
