+++ debian/control
@@ -1,9 +1,10 @@
 Source: coverxygen
-Section: utils
+Section: python
 Priority: optional
 Maintainer: Xavier MARCELET <xavier@marcelet.com>
-Build-Depends: debhelper (>= 7.0.50), python2.7 (>= 2.7), python3 (>= 3.4), python-setuptools, python3-setuptools
-X-Python3-Version: >= 3.4
+Standards-Version: 3.9.7
+Build-Depends: debhelper (>= 9), python-all (>= 2.7), python3-all (>= 3.2), python-setuptools, python3-setuptools
+X-Python3-Version: >= 3.2
 X-Python2-Version: >= 2.7
 
 Package: python-coverxygen
