+++ debian/source/options
@@ -1,0 +1,6 @@
+tar-ignore = ".git*"
+tar-ignore = ".travis.yml"
+tar-ignore = "build*"
+tar-ignore = ".pybuild*"
+tar-ignore = "coverxygen.egg-info"
+tar-ignore = "*__pycache__*"
